Added the song module files
