+++ app/build.gradle.kts
@@ -55,6 +55,7 @@ dependencies {
 
     implementation(projects.designSystem)
     implementation(projects.features.home.wiring)
+    implementation(projects.features.song.wiring)
 
     implementation(Dependencies.coreKtx)
     implementation(Dependencies.lifecycleRuntimeKtx)
