+++ features/home/impl/src/main/java/com/betocrod/features/song/impl/model/HomeState.kt
@@ -1,0 +1,9 @@
+package com.betocrod.features.song.impl.model
+
+import com.betocrod.features.song.impl.models.Song
+
+sealed class HomeState {
+
+    class SongsLoaded(val songs: List<Song>) : HomeState()
+    object Error : HomeState()
+}
