+++ features/home/impl/src/main/java/com/betocrod/features/song/impl/models/Song.kt
@@ -1,0 +1,8 @@
+package com.betocrod.features.song.impl.models
+
+data class Song(
+    val name: String,
+    val image: String,
+    val artist: String,
+    val year: Int
+)
