+++ features/song/api/build.gradle.kts
@@ -1,0 +1,13 @@
+plugins {
+    id("com.android.library")
+    id("org.jetbrains.kotlin.android")
+    id("com.betocrod.songrecords.android.library")
+}
+
+android {
+    namespace = "com.betocrod.features.song.api"
+}
+
+dependencies {
+
+}
