+++ features/song/impl/src/main/res/values/strings.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="feature_home_no_songs_placeholder">There are no songs</string>
+    <string name="feature_home_error">Unexpected error</string>
+    <string name="feature_home_title">Song Records</string>
+</resources>
