+++ features/song/wiring/build.gradle.kts
@@ -1,0 +1,14 @@
+plugins {
+    id("com.android.library")
+    id("org.jetbrains.kotlin.android")
+    id("com.betocrod.songrecords.android.library")
+}
+
+android {
+    namespace = "com.betocrod.songrecords.features.song.wiring"
+}
+
+dependencies {
+    api(projects.features.song.api)
+    implementation(projects.features.song.impl)
+}
